advisory-board-ai/backend/requirements.txt
# Core Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM - Using SQLite for simplicity in development
sqlalchemy==2.0.23
alembic==1.13.1

# Vector Store & Embeddings
chromadb==0.4.22
sentence-transformers==2.2.2

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI/LLM Integration
openai==1.3.0
groq==0.4.1
# align with transformers/tokenizers constraints (<0.18); 0.17.3 is available on PyPI
huggingface-hub==0.17.3
transformers==4.35.0
torch==2.1.0

# Utilities
python-dotenv==1.0.0
# python-multipart already listed above
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4

# API & Testing
requests==2.31.0
httpx==0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1

# WebSocket & Real-time
websockets==12.0